stub-fix: fixing tests
